+++ src/test/java/net/sf/jabref/gui/search/MatchesHighlighterTest.java
@@ -2,12 +2,14 @@ package net.sf.jabref.gui.search;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
+import net.sf.jabref.logic.search.SearchQueryHighlightObservable;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
 import java.util.Collections;
 import java.util.Optional;
 import java.util.function.Predicate;
+import java.util.regex.Pattern;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -22,14 +24,14 @@ public class MatchesHighlighterTest {
 
     @Test
     public void testHighlightWords() throws Exception {
-        assertEquals("", MatchesHighlighter.highlightWordsWithHTML("", Collections.emptyList()));
-        assertEquals("Einstein", MatchesHighlighter.highlightWordsWithHTML("Einstein", Collections.emptyList()));
-        assertEquals("<span style=\"background-color:#3399FF;\">Einstein</span>", MatchesHighlighter.highlightWordsWithHTML("Einstein", Collections.singletonList("Einstein")));
+        assertEquals("", MatchesHighlighter.highlightWordsWithHTML("", Optional.empty()));
+        assertEquals("Einstein", MatchesHighlighter.highlightWordsWithHTML("Einstein", Optional.empty()));
+        assertEquals("<span style=\"background-color:#3399FF;\">Einstein</span>", MatchesHighlighter.highlightWordsWithHTML("Einstein", Optional.of(Pattern.compile("Einstein"))));
     }
 
     @Test(expected = NullPointerException.class)
     public void testNullText() {
-        MatchesHighlighter.highlightWordsWithHTML(null, Collections.emptyList());
+        MatchesHighlighter.highlightWordsWithHTML(null, Optional.empty());
     }
 
     @Test(expected = NullPointerException.class)
@@ -39,12 +41,12 @@ public class MatchesHighlighterTest {
 
     @Test
     public void testNoWords() throws Exception {
-        assertEquals(Optional.empty(), MatchesHighlighter.getPatternForWords(Collections.emptyList(), true, true));
+        assertEquals(Optional.empty(), SearchQueryHighlightObservable.getPatternForWords(Collections.emptyList(), true, true));
     }
 
     @Test
     public void testPatternCaseInsensitive() throws Exception {
-        Predicate<String> predicate = MatchesHighlighter.getPatternForWords(Collections.singletonList("abc"), true, false).get().asPredicate();
+        Predicate<String> predicate = SearchQueryHighlightObservable.getPatternForWords(Collections.singletonList("abc"), true, false).get().asPredicate();
         assertTrue(predicate.test("abc"));
         assertTrue(predicate.test("ABC"));
         assertFalse(predicate.test("abd"));
@@ -53,7 +55,7 @@ public class MatchesHighlighterTest {
 
     @Test
     public void testPatternCaseSensitive() throws Exception {
-        Predicate<String> predicate = MatchesHighlighter.getPatternForWords(Collections.singletonList("abc"), true, true).get().asPredicate();
+        Predicate<String> predicate = SearchQueryHighlightObservable.getPatternForWords(Collections.singletonList("abc"), true, true).get().asPredicate();
         assertTrue(predicate.test("abc"));
         assertFalse(predicate.test("ABC"));
         assertFalse(predicate.test("abd"));