+++ CHANGELOG.md
@@ -28,6 +28,7 @@ to [sourceforge feature requests](https://sourceforge.net/p/jabref/features/) by
 - Write all field keys in lower case. No more camel casing of field names. E.g., `title` is written instead of `Title`, `howpublished` instead of `HowPublished`, and `doi` instead of `DOI`. The configuration option `Use camel case for field names (e.g., "HowPublished" instead of "howpublished")` is gone.
 - All field saving options are removed. There is no more customization of field sorting. '=' is now appended to the field key instead of its value. The intendation is aligned for an entry and not for the entire database. Entry names are written in title case format.
 - Entries are only reformatted if they were changed during a session. There is no more mandatory reformatting of the entire database on save.
+- Implements #565: Highlighting matches works now also for regular expressions in preview panel and entry editor
 
 ### Fixed
 - Fixed #479: Import works again